use vizia

+++ style/src/font/size.rs
@@ -1,6 +1,6 @@
-pub const XS: u16 = 14;
-pub const SM: u16 = 18;
-pub const MD: u16 = 23;
-pub const LG: u16 = 29;
-pub const XL: u16 = 37;
-pub const XXL: u16 = 47;
+pub const XS: f32 = 14.0;
+pub const SM: f32 = 18.0;
+pub const MD: f32 = 23.0;
+pub const LG: f32 = 29.0;
+pub const XL: f32 = 37.0;
+pub const XXL: f32 = 47.0;
